+++ src/main/java/com/ifgoiano/urt/projetocaoguia/projetocaoguiabackend/formularios/model/FormularioRequestDTO.java
@@ -1,4 +1,7 @@
 package com.ifgoiano.urt.projetocaoguia.projetocaoguiabackend.formularios.model;
 
-public record FormularioRequestDTO(String resposta) {
+public record FormularioRequestDTO(
+        CategoriaFormulario categoria,
+        String observacao
+) {
 }