+++ src/main/java/com/ifgoiano/urt/projetocaoguia/projetocaoguiabackend/usuarios/model/UsuarioResponseDTO.java
@@ -1,7 +1,7 @@
 package com.ifgoiano.urt.projetocaoguia.projetocaoguiabackend.usuarios.model;
 
-public record UsuarioResponseDTO(Long id, String nome, String email) {
+public record UsuarioResponseDTO(Long id, String nome, String email, String telefone) {
     public UsuarioResponseDTO(Usuario usuario) {
-        this(usuario.getId(), usuario.getNome(), usuario.getEmail());
+        this(usuario.getId(), usuario.getNome(), usuario.getEmail(), usuario.getTelefone());
     }
 }